requirements.txt
beautifulsoup4==4.6.0
chardet==2.3.0
Django==1.10.1
django-ipware==1.1.6
djangorestframework==3.4.7
ghdiff==0.4
Pillow==3.4.2
psycopg2==2.7.1
six==1.10.0
textile==2.3.5
whitenoise==3.3.0